Create a no-JS solution and replace jQuery with a React component.
For accessibility reasons it is required that gov services work without JS.
The original prototype used jQuery datatables and was entirely dependent on JS as
a result, with no backend whatsoever. As a proprietary solution it was also relatively
inflexible when adding or changing content.

This commit creates a Flask backend which stores the data (still as JSON) and uses
jinja2 templates for the frontend. This completely removes the need for JS. To
maintain the current user experience a React component has been added which replaces
the no JS solution if JS is available. The React component then controls the filtering
and also updates the URL to make it shareable.

The CSV and Excel exports now export using the current filter.

An open JSON API has also been introduced at /api/tariff.

The README has also been updated and a Circle CI configuration added.