Add metadata for use by data catalogues
This adds a metadata endpoint returning JSON that
conforms to the `data.json` format, as described:
https://guidance.data.gov.uk/publish_and_manage_data/harvest_or_add_data/harvest_data/dcat/

This file hardcodes the modified date so we'll
need to update this manually whenever we update
the data (which shouldn't be often).

Content for the `description` field has been
created from the GOV.UK content.