Fix(TP-93): Make filter dropdown more usable.
Currently the filter dropdown is supposed to work "intuitively" by
hitting enter when selecting the number of items to see on each page.
However this seems to only work on certain versions of certain
browsers.

To fix this a filter button is being made more obvious (in replacement
of the magnifying glass button). That way It is obvious what
should be clicked to cause the page to run the query.

Polyfills for IE and older versions of certain browsers have also
been added.