Improve performance.
Currently performance on a standard PaaS server starts to bottleneck
at around 6rps. This is a issue as the service is designed to be an
open service accessible to all, including an open API. There were some low hanging fruit to improve performance, but it was found that the majority of the bottleneck was due to Flask struggling to return more than ~30 MB/s.

This commit introduces several performance tweaks:

- Gevent workers are introduced in case the service become I/O bound.
- The JSON data is loaded into memory upfront so the file doesn't have to be continually interpreted.
- The filter step is skipped for any falsy param, not just None.
- All responses are now put through GZIP compression, significantly reducing the network overhead.
- After compression responses without arguments, or with default arguments, are cached in memory to save processing time.
- More memory has been given to the instance to give more leeway for the above.

Current testing shows this improves performance to approx.
30rps. Autoscaling will hopefully improve this further.