Add Google Analytics tracking to all views.
It is required that we use Google Analytics to track how the service
is used. To alay privacy concerns and avoid storing cookies this is
done on the backend instead of the frontend. The IP is sent to Google
however this is anonymised by Google so should not be an issue.

As this is an I/O operation that depends on an external service it
presents a performance risk. To handle this all google anayltics
requests are put into a thread to run separately from the main
request thread. This way the response can be sent to the client
without having to wait for the analytics request to be sent.