Add correct URL for the site header.
GDS have given a starting page to be hosted on gov.uk, we should direct to
this page when the header is clicked on. Similarly as GDS are hosting the
start page we should redirect / to the main tariff page - / is not being
reassigned as the main tariff page in case there are new features to host
down the line.

A small bug was also fixed (row values were assumed to be strings, this is
not always the case).