Add trade remedy information for specific rows.
In many cases there is a trade remedy which applies to a specific row.
As this introduces a level of complexity that can't be described in a single
table cell we have opted to introduce a popup which describes the remedy.
This is shown on click for JavaScript enabled browsers, and on hover for those
with JavaScript disabled.