Update data transform script to reflect latest structure.
The data we receive to inform this tool has changed structure. This commit
updates the transform script to reflect this. New columns have been added for
things such as trade remedies, suspensions, EPS and Meursing affected
commodities.

Some of the formatting has had to be refined as well (e.g. translating
floats to percentages).