Open the no-JS modal on :target.
Currently the no JS modal opens on :hover for the relevant cell.
This has the side effect of often making the URL within the modals
unclickable as the modal disappears as the mouse moves away from the
cell and towards the link.

By switching the See Detail span to an anchor for the cell (which
now has an ID) and showing the modal on :target this problem is solved.
The close button is then added back to the no JS modal to remove the
target from the URL, removing the modal.