Fix google analytics bugs.
As this tool depends heavily on query parameters it is important these
are tracked effectively. Likewise to insure we are accessible enough
it is important to track which languages use the tool the most. To
know what is driving traffic the Referer header must be tracked.
Finally it is equally important that if we can't send data to GA
that the response is still sent back to the client.

To cater for all this this commit firstly combines the path with the
query string. Secondly modifies how we extract the users language to
be in line with flasks system. The Referer header is forwarded to GA
if it exists. Lastly everything is wrapped in a try/except block to
ensure the response is always passed.