TP-155: Add an accessibility statement.
An accessibility statement is required to be added to the app.

This commit introduces the page as well as a link to the page in the footer.
There are also some minor CSS changes, mostly to improve the file format,
including a small change to the font size to improve readability on the
accessibility page.