Include DCAT-style XML metadata
Much of the metadata that data.gov.uk is capable of
showing just isn't picked up by their JSON importer.
So instead, switch to using the XML format so that
we can take full advantage of the metadata. We'll
remove the JSON metadata in a later commit, if this
is successful.