Replace "dumping margin" column with more accurate field.
Previously, we used the "Dumping margin" column to
decide which trade remedy modal to show: if the
"Dumping margin" column was TRUE, we'd show the
"tariff will apply from 1 Jan" modal, and if the
"Dumping margin" column was FALSE but the "Trade
remedy" was TRUE we'd show the "CET will apply
until transition reviews completed" modal.

We have discovered that "tariff will apply from
1 Jan" should apply in more circumstances than
just when there is a dumping margin – it includes
other margins as well. We have renamed the column
and updated some of the data values to reflect
this fact.